+++ src/main/java/kr/hhplus/be/ecommerce/EcommerceApplication.java
@@ -2,8 +2,12 @@ package kr.hhplus.be.ecommerce;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @SpringBootApplication
+@EnableJpaRepositories(basePackages = "kr.hhplus.be.ecommerce.infrastructure")
+@EntityScan(basePackages = "kr.hhplus.be.ecommerce.domain")
 public class EcommerceApplication {
 
 	public static void main(String[] args) {
