+++ src/test/java/kr/hhplus/be/ecommerce/ServerApplicationTests.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server;
+package kr.hhplus.be.ecommerce;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
