+++ src/test/java/kr/hhplus/be/ecommerce/TestcontainersConfiguration.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server;
+package kr.hhplus.be.ecommerce;
 
 import jakarta.annotation.PreDestroy;
 import org.springframework.boot.test.context.TestConfiguration;
